@@ -1,47 +1,49 @@
 {
-  "name": "unblock",
+  "name": "nextjs-subscription-payments",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
-    "dev": "next",
+    "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "stripe:listen": "stripe listen --forward-to=localhost:3000/api/webhooks --project-name=saas-starter"
+    "lint": "next lint",
+    "stripe:listen": "stripe listen --forward-to=localhost:3000/api/webhooks --project-name=saas-starter",
+    "generate-types": "npx supabase gen types typescript --project-id YOUR_PROJECT_ID --schema public > types_db.ts"
   },
   "dependencies": {
-    "@stripe/stripe-js": "^1.48.0",
-    "@supabase/auth-helpers-nextjs": "^0.5.4",
-    "@supabase/auth-helpers-react": "^0.3.1",
-    "@supabase/auth-ui-react": "^0.3.3",
-    "@supabase/auth-ui-shared": "^0.1.2",
-    "@supabase/supabase-js": "^2.10.0",
+    "@stripe/stripe-js": "^1.53.0",
+    "@supabase/auth-helpers-nextjs": "^0.7.0",
+    "@supabase/auth-ui-react": "^0.4.2",
+    "@supabase/auth-ui-shared": "^0.1.6",
+    "@supabase/supabase-js": "^2.23.0",
     "classnames": "^2.3.2",
-    "next": "^13.2.3",
+    "next": "13.4.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-merge-refs": "^2.0.1",
-    "stripe": "^11.18.0",
-    "swr": "^2.0.4",
-    "tailwindcss": "^3.2.7"
+    "react-merge-refs": "^2.0.2",
+    "stripe": "^12.6.0",
+    "tailwindcss": "^3.3.2"
   },
   "devDependencies": {
-    "@types/node": "^18.14.4",
-    "@types/react": "^18.0.28",
-    "autoprefixer": "^10.4.13",
-    "postcss": "^8.4.21",
-    "prettier": "^2.8.4",
-    "supabase": "^1.62.4",
-    "typescript": "^4.9.5"
+    "@ianvs/prettier-plugin-sort-imports": "^3.7.2",
+    "@types/node": "^20.2.3",
+    "@types/react": "^18.2.7",
+    "autoprefixer": "^10.4.14",
+    "eslint": "^8.41.0",
+    "eslint-config-next": "13.4.3",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-react": "^7.32.2",
+    "eslint-plugin-tailwindcss": "^3.12.0",
+    "postcss": "^8.4.23",
+    "prettier": "^2.8.8",
+    "prettier-plugin-tailwindcss": "^0.3.0",
+    "supabase": "^1.63.1",
+    "typescript": "^5.0.4"
   },
   "prettier": {
     "arrowParens": "always",
     "singleQuote": true,
     "tabWidth": 2,
     "trailingComma": "none"
-  },
-  "browser": {
-    "fs": false,
-    "os": false,
-    "path": false
   }
 }
