@@ -1,61 +1,62 @@
-// This code defines the Pricing component, which renders a page with pricing plans for a subscription service.
-
-// The Pricing component takes a single prop, products, which is an array of objects representing the pricing plans.
-
-// The Pricing component first checks if the products array is empty. If it is, the component renders a message saying that no pricing plans have been found.
-
-// If the products array is not empty, the component renders a section with the following elements:
-
-// * A header with the text "Pricing Plans"
-// * A paragraph with the text "Start building for free, then add a site plan to go live. Account plans unlock additional features."
-// * A set of buttons that allow the user to select a billing interval (month or year)
-// * A grid of cards, one for each pricing plan. Each card includes the following information:
-//* The name of the pricing plan
-//* The description of the pricing plan
-//* The price of the pricing plan
-//* A button that allows the user to subscribe to the pricing plan
-
-// The Pricing component also includes a footer with the logos of the following companies:
-
-// * Next.js
-// * Vercel
-// * Stripe
-// * Supabase
-// * GitHub
-
-import { useState } from 'react';
-import { useRouter } from 'next/router';
-import cn from 'classnames';
+'use client';
 
 import Button from '@/components/ui/Button';
+import { Database } from '@/types_db';
 import { postData } from '@/utils/helpers';
 import { getStripe } from '@/utils/stripe-client';
-import { useUser } from '@/utils/useUser';
+import { Session, User } from '@supabase/supabase-js';
+import cn from 'classnames';
+import { useRouter } from 'next/navigation';
+import { useState } from 'react';
 
-import { Price, ProductWithPrice } from 'types';
-
-interface Props {
-  products: ProductWithPrice[];
+type Subscription = Database['public']['Tables']['subscriptions']['Row'];
+type Product = Database['public']['Tables']['products']['Row'];
+type Price = Database['public']['Tables']['prices']['Row'];
+interface ProductWithPrices extends Product {
+  prices: Price[];
+}
+interface PriceWithProduct extends Price {
+  products: Product | null;
+}
+interface SubscriptionWithProduct extends Subscription {
+  prices: PriceWithProduct | null;
 }
 
-type BillingInterval = 'year' | 'month';
+interface Props {
+  session: Session | null;
+  user: User | null | undefined;
+  products: ProductWithPrices[];
+  subscription: SubscriptionWithProduct | null;
+}
 
-export default function Pricing({ products }: Props) {
+type BillingInterval = 'lifetime' | 'year' | 'month';
+
+export default function Pricing({
+  session,
+  user,
+  products,
+  subscription
+}: Props) {
+  const intervals = Array.from(
+    new Set(
+      products.flatMap((product) =>
+        product?.prices?.map((price) => price?.interval)
+      )
+    )
+  );
   const router = useRouter();
   const [billingInterval, setBillingInterval] =
     useState<BillingInterval>('month');
   const [priceIdLoading, setPriceIdLoading] = useState<string>();
-  const { user, isLoading, subscription } = useUser();
 
   const handleCheckout = async (price: Price) => {
     setPriceIdLoading(price.id);
     if (!user) {
       return router.push('/signin');
     }
-    if (subscription) {
+    if (price.product_id === subscription?.prices?.products?.id) {
       return router.push('/account');
     }
-
     try {
       const { sessionId } = await postData({
         url: '/api/create-checkout-session',
@@ -74,9 +75,9 @@ export default function Pricing({ products }: Props) {
   if (!products.length)
     return (
       <section className="bg-black">
-        <div className="max-w-6xl mx-auto py-8 sm:py-24 px-4 sm:px-6 lg:px-8">
+        <div className="max-w-6xl px-4 py-8 mx-auto sm:py-24 sm:px-6 lg:px-8">
           <div className="sm:flex sm:flex-col sm:align-center"></div>
-          <p className="text-6xl font-extrabold text-white sm:text-center sm:text-6xl">
+          <p className="text-4xl font-extrabold text-white sm:text-center sm:text-6xl">
             No subscription pricing plans found. Create them in your{' '}
             <a
               className="text-pink-500 underline"
@@ -89,46 +90,119 @@ export default function Pricing({ products }: Props) {
             .
           </p>
         </div>
+        <LogoCloud />
+      </section>
+    );
+
+  if (products.length === 1)
+    return (
+      <section className="bg-black">
+        <div className="max-w-6xl px-4 py-8 mx-auto sm:py-24 sm:px-6 lg:px-8">
+          <div className="sm:flex sm:flex-col sm:align-center">
+            <h1 className="text-4xl font-extrabold text-white sm:text-center sm:text-6xl">
+              Pricing Plans
+            </h1>
+            <p className="max-w-2xl m-auto mt-5 text-xl text-zinc-200 sm:text-center sm:text-2xl">
+              Start building for free, then add a site plan to go live. Account
+              plans unlock additional features.
+            </p>
+            <div className="relative flex self-center mt-12 border rounded-lg bg-zinc-900 border-zinc-800">
+              <div className="border border-pink-500 border-opacity-50 divide-y rounded-lg shadow-sm bg-zinc-900 divide-zinc-600">
+                <div className="p-6 py-2 m-1 text-2xl font-medium text-white rounded-md shadow-sm border-zinc-800 whitespace-nowrap focus:outline-none focus:ring-2 focus:ring-pink-500 focus:ring-opacity-50 focus:z-10 sm:w-auto sm:px-8">
+                  {products[0].name}
+                </div>
+              </div>
+            </div>
+            <div className="mt-6 space-y-4 sm:mt-12 sm:space-y-0 sm:grid sm:grid-cols-2 sm:gap-6 lg:max-w-4xl lg:mx-auto xl:max-w-none xl:mx-0 xl:grid-cols-3">
+              {products[0].prices?.map((price) => {
+                const priceString =
+                  price.unit_amount &&
+                  new Intl.NumberFormat('en-US', {
+                    style: 'currency',
+                    currency: price.currency!,
+                    minimumFractionDigits: 0
+                  }).format(price.unit_amount / 100);
+
+                return (
+                  <div
+                    key={price.interval}
+                    className="divide-y rounded-lg shadow-sm divide-zinc-600 bg-zinc-900"
+                  >
+                    <div className="p-6">
+                      <p>
+                        <span className="text-5xl font-extrabold white">
+                          {priceString}
+                        </span>
+                        <span className="text-base font-medium text-zinc-100">
+                          /{price.interval}
+                        </span>
+                      </p>
+                      <p className="mt-4 text-zinc-300">{price.description}</p>
+                      <Button
+                        variant="slim"
+                        type="button"
+                        disabled={false}
+                        loading={priceIdLoading === price.id}
+                        onClick={() => handleCheckout(price)}
+                        className="block w-full py-2 mt-12 text-sm font-semibold text-center text-white rounded-md hover:bg-zinc-900 "
+                      >
+                        {products[0].name ===
+                        subscription?.prices?.products?.name
+                          ? 'Manage'
+                          : 'Subscribe'}
+                      </Button>
+                    </div>
+                  </div>
+                );
+              })}
+            </div>
+          </div>
+          <LogoCloud />
+        </div>
       </section>
     );
 
   return (
     <section className="bg-black">
-      <div className="max-w-6xl mx-auto py-8 sm:py-24 px-4 sm:px-6 lg:px-8">
+      <div className="max-w-6xl px-4 py-8 mx-auto sm:py-24 sm:px-6 lg:px-8">
         <div className="sm:flex sm:flex-col sm:align-center">
           <h1 className="text-4xl font-extrabold text-white sm:text-center sm:text-6xl">
             Pricing Plans
           </h1>
-          <p className="mt-5 text-xl text-zinc-200 sm:text-center sm:text-2xl max-w-2xl m-auto">
+          <p className="max-w-2xl m-auto mt-5 text-xl text-zinc-200 sm:text-center sm:text-2xl">
             Start building for free, then add a site plan to go live. Account
             plans unlock additional features.
           </p>
           <div className="relative self-center mt-6 bg-zinc-900 rounded-lg p-0.5 flex sm:mt-8 border border-zinc-800">
-            <button
-              onClick={() => setBillingInterval('month')}
-              type="button"
-              className={`${
-                billingInterval === 'month'
-                  ? 'relative w-1/2 bg-zinc-700 border-zinc-800 shadow-sm text-white'
-                  : 'ml-0.5 relative w-1/2 border border-transparent text-zinc-400'
-              } rounded-md m-1 py-2 text-sm font-medium whitespace-nowrap focus:outline-none focus:ring-2 focus:ring-pink-500 focus:ring-opacity-50 focus:z-10 sm:w-auto sm:px-8`}
-            >
-              Monthly billing
-            </button>
-            <button
-              onClick={() => setBillingInterval('year')}
-              type="button"
-              className={`${
-                billingInterval === 'year'
-                  ? 'relative w-1/2 bg-zinc-700 border-zinc-800 shadow-sm text-white'
-                  : 'ml-0.5 relative w-1/2 border border-transparent text-zinc-400'
-              } rounded-md m-1 py-2 text-sm font-medium whitespace-nowrap focus:outline-none focus:ring-2 focus:ring-pink-500 focus:ring-opacity-50 focus:z-10 sm:w-auto sm:px-8`}
-            >
-              Yearly billing
-            </button>
+            {intervals.includes('month') && (
+              <button
+                onClick={() => setBillingInterval('month')}
+                type="button"
+                className={`${
+                  billingInterval === 'month'
+                    ? 'relative w-1/2 bg-zinc-700 border-zinc-800 shadow-sm text-white'
+                    : 'ml-0.5 relative w-1/2 border border-transparent text-zinc-400'
+                } rounded-md m-1 py-2 text-sm font-medium whitespace-nowrap focus:outline-none focus:ring-2 focus:ring-pink-500 focus:ring-opacity-50 focus:z-10 sm:w-auto sm:px-8`}
+              >
+                Monthly billing
+              </button>
+            )}
+            {intervals.includes('year') && (
+              <button
+                onClick={() => setBillingInterval('year')}
+                type="button"
+                className={`${
+                  billingInterval === 'year'
+                    ? 'relative w-1/2 bg-zinc-700 border-zinc-800 shadow-sm text-white'
+                    : 'ml-0.5 relative w-1/2 border border-transparent text-zinc-400'
+                } rounded-md m-1 py-2 text-sm font-medium whitespace-nowrap focus:outline-none focus:ring-2 focus:ring-pink-500 focus:ring-opacity-50 focus:z-10 sm:w-auto sm:px-8`}
+              >
+                Yearly billing
+              </button>
+            )}
           </div>
         </div>
-        <div className="mt-12 space-y-4 sm:mt-16 sm:space-y-0 sm:grid sm:grid-cols-2 sm:gap-6 lg:max-w-4xl lg:mx-auto xl:max-w-none xl:mx-0 xl:grid-cols-4">
+        <div className="mt-12 space-y-4 sm:mt-16 sm:space-y-0 sm:grid sm:grid-cols-2 sm:gap-6 lg:max-w-4xl lg:mx-auto xl:max-w-none xl:mx-0 xl:grid-cols-3">
           {products.map((product) => {
             const price = product?.prices?.find(
               (price) => price.interval === billingInterval
@@ -136,7 +210,7 @@ export default function Pricing({ products }: Props) {
             if (!price) return null;
             const priceString = new Intl.NumberFormat('en-US', {
               style: 'currency',
-              currency: price.currency,
+              currency: price.currency!,
               minimumFractionDigits: 0
             }).format((price?.unit_amount || 0) / 100);
             return (
@@ -152,7 +226,7 @@ export default function Pricing({ products }: Props) {
                 )}
               >
                 <div className="p-6">
-                  <h2 className="text-2xl leading-6 font-semibold text-white">
+                  <h2 className="text-2xl font-semibold leading-6 text-white">
                     {product.name}
                   </h2>
                   <p className="mt-4 text-zinc-300">{product.description}</p>
@@ -167,10 +241,10 @@ export default function Pricing({ products }: Props) {
                   <Button
                     variant="slim"
                     type="button"
-                    disabled={isLoading}
+                    disabled={!session}
                     loading={priceIdLoading === price.id}
                     onClick={() => handleCheckout(price)}
-                    className="mt-8 block w-full rounded-md py-2 text-sm font-semibold text-white text-center hover:bg-zinc-900"
+                    className="block w-full py-2 mt-8 text-sm font-semibold text-center text-white rounded-md hover:bg-zinc-900"
                   >
                     {product.name === subscription?.prices?.products?.name
                       ? 'Manage'
@@ -181,59 +255,65 @@ export default function Pricing({ products }: Props) {
             );
           })}
         </div>
-        <div>
-          <p className="mt-24 text-xs uppercase text-zinc-400 text-center font-bold tracking-[0.3em]">
-            Brought to you by
-          </p>
-          <div className="flex flex-col items-center my-12 space-y-4 sm:mt-8 sm:space-y-0 md:mx-auto md:max-w-2xl sm:grid sm:gap-6 sm:grid-cols-5">
-            <div className="flex items-center justify-start">
-              <a href="https://nextjs.org" aria-label="Next.js Link">
-                <img
-                  src="/nextjs.svg"
-                  alt="Next.js Logo"
-                  className="h-12 text-white"
-                />
-              </a>
-            </div>
-            <div className="flex items-center justify-start">
-              <a href="https://vercel.com" aria-label="Vercel.com Link">
-                <img
-                  src="/vercel.svg"
-                  alt="Vercel.com Logo"
-                  className="h-6 text-white"
-                />
-              </a>
-            </div>
-            <div className="flex items-center justify-start">
-              <a href="https://stripe.com" aria-label="stripe.com Link">
-                <img
-                  src="/stripe.svg"
-                  alt="stripe.com Logo"
-                  className="h-12 text-white"
-                />
-              </a>
-            </div>
-            <div className="flex items-center justify-start">
-              <a href="https://supabase.io" aria-label="supabase.io Link">
-                <img
-                  src="/supabase.svg"
-                  alt="supabase.io Logo"
-                  className="h-10 text-white"
-                />
-              </a>
-            </div>
-            <div className="flex items-center justify-start">
-              <a href="https://github.com" aria-label="github.com Link">
-                <img
-                  src="/github.svg"
-                  alt="github.com Logo"
-                  className="h-8 text-white"
-                />
-              </a>
-            </div>
-          </div>
-        </div>
+        <LogoCloud />
       </div>
     </section>
+  );
+}
+
+function LogoCloud() {
+  return (
+    <div>
+      <p className="mt-24 text-xs uppercase text-zinc-400 text-center font-bold tracking-[0.3em]">
+        Brought to you by
+      </p>
+      <div className="flex flex-col items-center my-12 space-y-4 sm:mt-8 sm:space-y-0 md:mx-auto md:max-w-2xl sm:grid sm:gap-6 sm:grid-cols-5">
+        <div className="flex items-center justify-start">
+          <a href="https://nextjs.org" aria-label="Next.js Link">
+            <img
+              src="/nextjs.svg"
+              alt="Next.js Logo"
+              className="h-12 text-white"
+            />
+          </a>
+        </div>
+        <div className="flex items-center justify-start">
+          <a href="https://vercel.com" aria-label="Vercel.com Link">
+            <img
+              src="/vercel.svg"
+              alt="Vercel.com Logo"
+              className="h-6 text-white"
+            />
+          </a>
+        </div>
+        <div className="flex items-center justify-start">
+          <a href="https://stripe.com" aria-label="stripe.com Link">
+            <img
+              src="/stripe.svg"
+              alt="stripe.com Logo"
+              className="h-12 text-white"
+            />
+          </a>
+        </div>
+        <div className="flex items-center justify-start">
+          <a href="https://supabase.io" aria-label="supabase.io Link">
+            <img
+              src="/supabase.svg"
+              alt="supabase.io Logo"
+              className="h-10 text-white"
+            />
+          </a>
+        </div>
+        <div className="flex items-center justify-start">
+          <a href="https://github.com" aria-label="github.com Link">
+            <img
+              src="/github.svg"
+              alt="github.com Logo"
+              className="h-8 text-white"
+            />
+          </a>
+        </div>
+      </div>
+    </div>
   );
 }
